requirements.txt
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
click==8.0.3
# pip --trusted-host=pypi.python.org --trusted-host=pypi.org --trusted-host=files.pythonhosted.org install pandas --user
pandas
pycryptodome